moved conditional logic into Game instead of App
