+++ src/components/Game.js
@@ -3,17 +3,38 @@ import { firestoreConnect } from 'react-redux-firebase';
 import { compose } from 'redux';
 import { withRouter } from 'react-router-dom';
 import { SidebarCards, Board, ActionFooter } from './index';
+import { db } from '../store';
 import history from '../history';
 
-const Game = () => {
-  return (
-    <div className="game">
-      <SidebarCards />
-      <Board />
-      <ActionFooter />
-    </div>
-  );
-};
+class Game extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+      gameExists: false
+    };
+  }
+
+  async componentWillMount() {
+    const gameId = history.location.pathname.slice(1);
+    const game = gameId && await db.collection('games').doc(gameId).get();
+    if (game.exists) {
+      this.setState({ gameExists: true });
+    } else {
+      history.push('/');
+    }
+  }
+
+  render() {
+    return (
+      this.state.gameExists &&
+        <div className="game">
+          <SidebarCards />
+          <Board />
+          <ActionFooter />
+        </div>
+    );
+  }
+}
 
 export default withRouter(compose(
   firestoreConnect(() => [
